chore(sdk): generate go extend sdk (230605.2233)
- generated from source sdk spec commit: 804dfcf8d7eeb471ee584df4a90acb1e12a6c6ff

+++ dsmc-sdk/pkg/dsmcclientmodels/models_create_image_request.go
@@ -32,8 +32,8 @@ type ModelsCreateImageRequest struct {
 
 	// imagesize
 	// Required: true
-	// Format: int32
-	ImageSize *int32 `json:"imageSize"`
+	// Format: int64
+	ImageSize *int64 `json:"imageSize"`
 
 	// namespace
 	// Required: true
