+++ dsmc-sdk/pkg/dsmcclientmodels/models_image_record.go
@@ -37,8 +37,8 @@ type ModelsImageRecord struct {
 
 	// imagesize
 	// Required: true
-	// Format: int32
-	ImageSize *int32 `json:"imageSize"`
+	// Format: int64
+	ImageSize *int64 `json:"imageSize"`
 
 	// modifiedby
 	// Required: true
